requirements.txt
azure-cosmos==4.5.0
pandas==2.1.0
numpy~=1.25.2
fastapi==0.103.0
matplotlib
scikit-learn
plotly==5.16.1
openai==0.27.10
pydantic==2.3.0
uvicorn
python-dotenv~=1.0.0